requirements.txt
pyguitarpro==0.9.3
